[#97] refactor : 로그인여부, 관리자 권한 확인  custom hook으로 리펙토링
